refactor design to work with a single page of html to simplify preservation of dynamic game state
Initially, I wanted to reuse parts of my UI design from another project (hw2 - portfolio)
This included 3 links to static html content.

However, here, the content is dynamic since game state is reflected in
dynamically created js objects.  Sadly, moving from the Play page to the Stats page,
for example, had the side-effect of entirely resetting the game state.

In a more sophisticated implementation, game state could probably be managed on the server
(perhaps in combination with cookies or local storage).

However in this case, with a client-side implementation, the simplest solution is to have a
single page of html from which the js runs and reimplement Stats and
Help as simple alert boxes.