add game play logic
This is /mostly/ working.  I need to clean-up an odd race condition
whereby the win/loss alert appears on the screen before the DOM
has finished updating the guessed word text and the guess remaining.

:-/

Also, it would be good to unregister the keyboard event handler
clean after a win/loss.  Then wait for the user to press play
again, if they want to continue.