fix regression when disabling 'Play' button
I forgot to disable the event listener I registered to the now-nonexistent
Play button.  This ended up killing the downstream registration of
other keyboard event handlers.  Consequently no keyboard events
were generated when the user typed in letters.

Also, on smaller form factors, I'm taking the hidden button entirely
out of the rendering flow with 'display: none' so that it doesn't
distort the centering of the other menu links.