wire up event listener to theme selection drop-down
The theme-change feature is finally live with this commit.

We let the default model lexicon theme drive the initial selected option
and then we listen for selection change events to trigger a
change of lexicons in the model, followed by a controller and game
reset.

We still need to update the background images associated with each themed
lexicon.  Maybe spruce the word lists too.  But that is all icing on the cake.